Split tests into two modules
